requirements.txt
Flask
Flask-SQLAlchemy
Flask-SocketIO
psycopg2-binary
Werkzeug
Flask-SQLAlchemy
SQLAlchemy
Flask-SocketIO